All .md files: content review and re-organisation, v3
Started on 28th July 2023, this is the commit that squashes all the commits
related to the topic above and is going to set the first version of this
repository. Every .md page content before this commit should be considered
just notes taken during the 4+4 weeks preliminary SFOS analysis, which brought
to the SFOS 4-stages refactoring plan code-named RedFish OS (RFOS).

This .md pages content refactoring was completed on 31th July 2023.

Signed-off-by: Roberto A. Foglietta <roberto.foglietta@gmail.com>